Fixed a little bug
If you pressed the button to delete a specific player out of the
buddies, but you didn't select anybody, then it created an error. This
bug is fixed now.

+++ lua/patchprotect/client/buddy.lua
@@ -100,6 +100,8 @@ end
 
 -- DELETE BUDDY
 function cl_PProtect.DeleteBuddy( ply )
+
+	if !ply then return end
 	
 	ply.Buddies = ply.Buddies or {}
 	sql.Query( "DELETE FROM pprotect_buddies WHERE uniqueid = '" .. cl_PProtect.Buddy.BuddyToRemove[0] .. "'" )
